requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.3.0
werkzeug==3.0.3
zipp==3.19.2
